Corrige install e gemrc

+++ install.sh
@@ -143,6 +143,11 @@ if ! command_exists node; then
 fi
 
 
-echo -n "[ Ruby -> Bundler - Solargraph ]"
-gem install bundler
-gem install solargraph
+if command_exists ruby; then
+  echo -n "[ Ruby -> Bundler - Solargraph ]"
+  gem install bundler
+  gem install solargraph
+else
+  rbenv install 2.6.6
+  rbenv global 2.6.6
+fi
